# auth-service/src/main/java/com/example/auth/adapters/in/web/dto/LoginRequest.java
package com.example.auth.adapters.in.web.dto;

import jakarta.validation.constraints.NotBlank;

public class LoginRequest {

    @NotBlank(message = "{error.username.required}")
    private String username;

    @NotBlank(message = "{error.password.required}")
    private String password;

    public LoginRequest() {}

    public LoginRequest(String username, String password) {
        this.username = username;
        this.password = password;
    }

    public String getUsername() { return username; }
    public String getPassword() { return password; }
    public void setUsername(String username) { this.username = username; }
    public void setPassword(String password) { this.password = password; }
}